+++ ShoppingCart/src/main/java/com/mindtreeyorbitsb201/shoppingcart/service/CartService.java
@@ -3,10 +3,10 @@ package com.mindtreeyorbitsb201.shoppingcart.service;
 import com.mindtreeyorbitsb201.shoppingcart.entity.Cart;
 
 public interface CartService {
-	Cart addProducts(Cart cart,Long userId,Long productId);
+	Cart addProducts(Cart cart,Integer userId,Integer productId);
 	Cart updateProducts(Cart cart);
-	void removeProducts(Long cartId,Long productId);
-	void removeAllProducts(Long cartId);
-	Cart viewProducts(Long userId);
+	void removeProducts(Integer cartId,Integer productId);
+	void removeAllProducts(Integer cartId);
+	Cart viewProducts(Integer userId);
 	
 }
